Fixed a bug that caused objects to be incorrectly culled at small fields of view. Changed type of Renderer::cosViewConeAngle from float to double; using single precision floats was causing the value of cosViewConeAngle to be rounded to 1.0, making the view cone effectively just a line.

+++ trunk/celestia/src/celengine/render.h
@@ -654,7 +654,7 @@ class Renderer
     int windowWidth;
     int windowHeight;
     float fov;
-    float cosViewConeAngle;
+    double cosViewConeAngle;
     int screenDpi;
     float corrFac;
     float pixelSize;
